Extract embed methods to EmbedFactory module
