+++ cogs/logging.py
@@ -3,7 +3,7 @@ import os
 from discord import TextChannel
 from discord.ext import commands
 
-from cogs.moderation import footer_embed
+from EmbedFactory import footer_embed
 
 
 class Logging(commands.Cog):
